Finish removing pytato.partition
