+++ doc/dag.rst
@@ -25,11 +25,6 @@ Stringifying Expression Graphs
 
 .. _partitioning:
 
-Partitioning Array Expression Graphs
-====================================
-
-.. automodule:: pytato.partition
-
 .. _distributed:
 
 Support for Distributed-Memory/Message Passing
